EVG-16337 move some queries to unclutter tasks
